third_party/blink/web_tests/svg/dynamic-updates/SVGClipPathElement-dom-clipPathUnits-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'clipPathUnits' attribute of the SVGClipPathElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS clipPathElement.getAttribute('clipPathUnits') is "userSpaceOnUse"
PASS clipPathElement.getAttribute('clipPathUnits') is "objectBoundingBox"
PASS successfullyParsed is true

TEST COMPLETE
